Add Lexer struct and New()
The input and char fields should be self-explanatory. The pos and readPos represent a standard two-pointer for examining the input which ensures the lexer is able to peek ahead into the input. This is useful to determine if the lexer is currently reading a keyword or identifier, which will span multiple characters, or a single character (which is marked by readPos pointing to a whitespace/newline).

The New() function creates the lexer, and is currently bare-bones. This will be expanded upon in the future.