Refactor lexer's NextToken for enhanced readability and maintainability
By introducing helper functions that handle token generation, we improve maintainability of the function. Consider when we introduce more complex symbols like '<=', we would bloat the original switch statement, and make it harder to read. By encapsulating this logic to a single function, the switch statement can be focused on symbols which should aid readability somewhat.

The logic for identifiers and numbers are removed from the original switch statement too. This aids readability by keeping the defualt case in the new switch statement in createTokenFromSymbol() simple.